restructure pkg hierarchy to fit to schema migrations
