--- central/backgroundmigrations/all.go
@@ -1,6 +1,0 @@
-package backgroundmigrations
-
-// Import background migration packages here to register them via init().
-// import (
-// 	_ "github.com/stackrox/rox/central/backgroundmigrations/m_0001_example"
-// )
