+++ central/backgroundmigrations/runner/all.go
@@ -1,0 +1,6 @@
+package runner
+
+import (
+	_ "github.com/stackrox/rox/central/backgroundmigrations/migrations/m_000_to_m_001_test_migration"
+	_ "github.com/stackrox/rox/central/backgroundmigrations/migrations/m_001_to_m_002_slow_migration"
+)
