+++ central/backgroundmigrations/types/migration.go
@@ -1,4 +1,4 @@
-package backgroundmigrations
+package types
 
 import (
 	"context"
